+++ docs/changelog.rst
@@ -27,6 +27,8 @@ Fixes
 
 * Recipe containing a MultiplotTask does not contain datasets as dicts
 * PlotTask with automatically generated filenames and >1 datasets writes correct filenames to figure record in recipe
+* CompositePlotter sets plot style of plotters
+
 
 Version 0.6.1
 =============